kisan_sathi_backend/weather/views.py
from decouple import config
import requests
from datetime import datetime
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status


class WeatherSummaryView(APIView):
    def get(self, request):
        city = request.query_params.get('q') or request.query_params.get('city') or 'Bengaluru'
        api_key = config('OPENWEATHER_API_KEY', default='')
        if not api_key or api_key == 'your-api-key':
            return Response({
                'success': False,
                'message': 'OPENWEATHER_API_KEY missing. Set it in backend .env and restart.'
            }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)

        try:
            current_url = 'https://api.openweathermap.org/data/2.5/weather'
            forecast_url = 'https://api.openweathermap.org/data/2.5/forecast'
            params = {'q': city, 'appid': api_key, 'units': 'metric'}

            current = requests.get(current_url, params=params, timeout=10)
            current.raise_for_status()
            current_data = current.json()

            forecast = requests.get(forecast_url, params=params, timeout=10)
            forecast.raise_for_status()
            forecast_data = forecast.json()

            # Prefer normalized city name from API (includes country)
            normalized_city = current_data.get('name')
            country = (current_data.get('sys') or {}).get('country')
            display_city = f"{normalized_city}, {country}" if normalized_city and country else (normalized_city or city)

            # Include current icon as well
            cur_icon = None
            try:
                cur_icon = (current_data.get('weather') or [{}])[0].get('icon')
            except Exception:
                cur_icon = None

            current_payload = {
                'temp': round(current_data['main']['temp']),
                'condition': current_data['weather'][0]['description'].title(),
                'humidity': current_data['main']['humidity'],
                'windSpeed': round(current_data['wind'].get('speed', 0)),
                'rainfall': round(current_data.get('rain', {}).get('1h', 0)),
                'icon': map_icon(cur_icon, current_data['weather'][0]['main']) if 'map_icon' in locals() else None,
                'iconUrl': f"https://openweathermap.org/img/wn/{cur_icon}@2x.png" if cur_icon else None,
            }

            # Helper: map OpenWeather icon code/main to a simple emoji
            def map_icon(icon_code: str, main: str) -> str:
                try:
                    if icon_code:
                        key = icon_code[:2]
                        return {
                            '01': '☀️',  # clear sky
                            '02': '🌤️',  # few clouds
                            '03': '☁️',  # scattered clouds
                            '04': '☁️',  # broken clouds
                            '09': '🌧️',  # shower rain
                            '10': '🌦️',  # rain
                            '11': '⛈️',  # thunderstorm
                            '13': '❄️',  # snow
                            '50': '🌫️',  # mist
                        }.get(key, '🌤️')
                    # fallback by main
                    return {
                        'Clear': '☀️',
                        'Clouds': '☁️',
                        'Rain': '🌦️',
                        'Drizzle': '🌧️',
                        'Thunderstorm': '⛈️',
                        'Snow': '❄️',
                        'Mist': '🌫️',
                        'Smoke': '🌫️',
                        'Haze': '🌫️',
                        'Dust': '🌫️',
                        'Fog': '🌫️',
                        'Sand': '🌫️',
                        'Ash': '🌫️',
                        'Squall': '🌬️',
                        'Tornado': '🌪️',
                    }.get(main, '🌤️')
                except Exception:
                    return '🌤️'

            # Build simple 5-day forecast from 3-hourly data (pick noon entries)
            days_seen = set()
            daily = []
            for item in forecast_data.get('list', []):
                dt = datetime.fromtimestamp(item['dt'])
                day_key = dt.strftime('%Y-%m-%d')
                hour = dt.hour
                if day_key not in days_seen and 11 <= hour <= 14:
                    days_seen.add(day_key)
                    ow_icon = None
                    try:
                        ow_icon = (item.get('weather') or [{}])[0].get('icon')
                    except Exception:
                        ow_icon = None
                    main = (item.get('weather') or [{}])[0].get('main', '')
                    ow_code = ow_icon or ''
                    daily.append({
                        'day': dt.strftime('%a'),
                        'high': round(item['main']['temp_max']),
                        'low': round(item['main']['temp_min']),
                        'condition': item['weather'][0]['main'],
                        'icon': map_icon(ow_icon, main),
                        'iconUrl': f"https://openweathermap.org/img/wn/{ow_code}@2x.png" if ow_code else None,
                        'rainfall': round(item.get('rain', {}).get('3h', 0)),
                    })
                if len(daily) >= 5:
                    break

            # If forecast payload has city info, use it
            try:
                fc_city_name = (forecast_data.get('city') or {}).get('name')
                fc_country = (forecast_data.get('city') or {}).get('country')
                if fc_city_name and fc_country:
                    display_city = f"{fc_city_name}, {fc_country}"
            except Exception:
                pass

            return Response({
                'success': True,
                'city': display_city,
                'current': current_payload,
                'forecast': daily,
                'alerts': [],
            })
        except requests.HTTPError as e:
            try:
                err = e.response.json()
            except Exception:
                err = {'message': str(e)}
            return Response({'success': False, 'message': 'OpenWeather error', 'error': err}, status=e.response.status_code)
        except Exception as e:
            # Graceful fallback when network/DNS fails
            fallback = {
                'success': True,
                'city': city,
                'current': {
                    'temp': 28,
                    'condition': 'Partly Cloudy',
                    'humidity': 65,
                    'windSpeed': 12,
                    'rainfall': 0,
                },
                'forecast': [
                    {'day': 'Tomorrow', 'high': 32, 'low': 22, 'condition': 'Sunny', 'icon': '☀️', 'rainfall': 0},
                    {'day': 'Day After', 'high': 30, 'low': 20, 'condition': 'Rainy', 'icon': '🌦️', 'rainfall': 25},
                    {'day': '3 Days', 'high': 28, 'low': 19, 'condition': 'Cloudy', 'icon': '☁️', 'rainfall': 10},
                    {'day': '4 Days', 'high': 31, 'low': 21, 'condition': 'Sunny', 'icon': '☀️', 'rainfall': 0},
                    {'day': '5 Days', 'high': 33, 'low': 23, 'condition': 'Hot', 'icon': '🔥', 'rainfall': 0},
                ],
                'alerts': [],
                'note': 'Using fallback data due to network error',
            }
            return Response(fallback, status=200)
