kisan_sathi_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
requests==2.31.0
celery==5.3.4
redis==5.0.1
